Leave just one destructurer based on expression trees
